pyproject.toml
[project]
name = "depth"
version = "0.1.0"
description = "Library for unsupervised depth estimation"
readme = "README.md"
authors = [
    { name = "Angus Stewart", email = "siliconlad@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "tinygrad>=0.12.0",
    "track-sdk>=0.2.0",
    "pybag-sdk>=0.13.0",
    "pillow",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
addopts = "-m 'not camera'"
markers = ["camera: requires ZED camera hardware"]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
